Factor out loopback AMP client.
